Create README.md
This is the initial readme file. Extra description to the project with my intentions. Some of the game elements are still a little convoluted. The vault door doesn't have proper collision points etc.. However the game does work as intended. And is in relation the the breakout game completion of 100D Bootcamp.